+++ CHANGELOG.md
@@ -2,13 +2,18 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+
+### Removed
+- Auth service, user registration, login, and account UI
+- Guest check limit that depended on authentication
+
 ## [1.0.0] - 2026-06-21
 
 ### Added
 - Issue templates
 - User-story index
 - (Core feature) Checking website for complience with the law FL-152
-- User registration and login
 
 ### Changed
 - Some user stories